removed AreaChartBody, updated yScaleDef function to work with AreaChart, modularized code for Area Chart
