+++ src/charts/AreaChart/AreaChart.tsx
@@ -1,9 +1,19 @@
 /** App.js */
-import React, { useState, useEffect, useRef } from "react";
-import useEnvEffect from '../../hooks/useEnvEffect';
+import React, { useMemo} from "react";
 import * as d3 from "d3";
-import AreaChartBody from "./AreaChartBody";
-import { AreaChartProps } from '../../../types';
+import { AreaChartProps, ColorScale, xAccessorFunc, yAccessorFunc } from '../../../types';
+import ContinuousAxis from "../../components/ContinuousAxis";
+import { useResponsive } from '../../hooks/useResponsive';
+import { xScaleDef } from '../../functionality/xScale';
+import { yScaleDef } from '../../functionality/yScale';
+import {
+  getXAxisCoordinates,
+  getYAxisCoordinates,
+  getMargins,
+  inferXDataType,
+  transformSkinnyToWide
+} from "../../utils";
+
 
 
 export default function AreaChart({
@@ -23,27 +33,71 @@ export default function AreaChart({
   colorScheme = d3.schemeCategory10
 }:AreaChartProps<string | number>):JSX.Element {
 
-  const anchor = useRef(null as unknown as SVGSVGElement);
-  const [windowSize, setWindowSize] = useState<[number, number]>([0, 0]);
-  const [cHeight, setCHeight] = useState<number>(0);
-  const [cWidth, setCWidth] = useState<number>(0);
+  const chart = 'AreaChart';
 
-  function updateSize() {
-    setWindowSize([window.innerWidth, window.innerHeight]);
+  const {anchor, cHeight, cWidth}  = useResponsive();
+  const margin = useMemo(
+    () => getMargins(xAxis, yAxis, xAxisLabel, yAxisLabel),
+    [xAxis, yAxis, xAxisLabel, yAxisLabel]
+  )
+  const { xAxisX, xAxisY } = useMemo(
+    () => getXAxisCoordinates(xAxis, cHeight, margin),
+    [cHeight, xAxis, margin]
+  )
+  const { yAxisX, yAxisY } = useMemo(
+    () => getYAxisCoordinates(yAxis, cWidth, margin),
+    [cWidth, yAxis, margin]
+  )
+  // offset group to match position of axes
+  const translate = `translate(${margin.left}, ${margin.top})`
+  
+  // type KeyType = { key: string; dataType?: "number" | "date" | undefined; }
+  
+  // if no xKey datatype is passed in, determine if it's Date
+  if (!xDataType) {
+    xDataType = inferXDataType(data[0], xKey);
   }
 
-  // Set up an event listener on mount
-  useEnvEffect(() => {
-    window.addEventListener("resize", updateSize);
-    updateSize();
-    return () => window.removeEventListener("resize", updateSize);
-  }, []);
+  // generate arr of keys. these are used to render discrete areas to be displayed
+  const keys: string[] = [];
+  if (groupBy) {
+    for (const entry of data) {
+      const property = String(entry[groupBy ?? ""]);
+      if (property && !keys.includes(property)) {
+        keys.push(property)
+      }
+    }
+    data = transformSkinnyToWide(data, keys, groupBy, xKey, yKey);
+  } else {
+    keys.push(yKey)
+  }
 
-  useEffect(() => {
-    const container = anchor.current.getBoundingClientRect();
-    setCHeight(container.height);
-    setCWidth(container.width);
-  }, [windowSize]);
+  // generate stack: an array of Series representing the x and associated y0 & y1 values for each area
+  const stack = d3.stack().keys(keys)
+  const layers = useMemo(() => {
+    const layersTemp = stack(data as Iterable<{ [key: string]: number; }>);
+    for (const series of layersTemp) {
+      series.sort((a, b) => b.data[xKey] - a.data[xKey]);
+    }
+    return layersTemp;
+  }, [data, keys])
+  
+  const xAccessor: xAccessorFunc = xDataType === 'number' ? (d) => d[xKey] : (d) => new Date(d[xKey]);
+  const yAccessor: yAccessorFunc = (d) => d[yKey];
+  
+  const {xScale, xMin, xMax} = xScaleDef(data, xDataType, xAccessor, margin, cWidth, chart);
+  const yScale = yScaleDef(layers, yAccessor, margin, cHeight, chart);   
+  
+  let xTicksValue = [xMin, ... xScale.ticks(), xMax]
+
+  const areaGenerator: any = d3
+    .area()
+    .x((layer: any) => xScale(xAccessor(layer.data)))
+    .y0((layer) => yScale(layer[0]))
+    .y1((layer) => yScale(layer[1]))
+
+  const colorScale: ColorScale = d3.scaleOrdinal(colorScheme)
+  colorScale.domain(keys)
 
   // IF data is structured the wrong way, convert:
   // let dataTransformed;
@@ -51,22 +105,42 @@ export default function AreaChart({
 
   return (
       <svg ref={anchor} width={width} height={height}>
-        <AreaChartBody
-          height={cHeight}
-          width={cWidth}
-          data={data}
-          xKey={xKey}
-          xDataType={xDataType}
-          yKey={yKey}
-          groupBy={groupBy}
-          xAxis={xAxis}
-          yAxis={yAxis}
-          xGrid={xGrid}
-          yGrid={yGrid}
-          xAxisLabel={xAxisLabel}
-          yAxisLabel={yAxisLabel}
-          colorScheme={colorScheme}
+         <g transform={translate}>
+      {yAxis && (
+        <ContinuousAxis
+        x={yAxisX}
+        y={yAxisY}
+        height={cHeight}
+        width={cWidth}
+        margin={margin}
+        scale={yScale}
+        type={yAxis}
+        yGrid={yGrid}
+        label={yAxisLabel}
         />
+      )}
+      {xAxis && (
+        <ContinuousAxis
+        x={xAxisX}
+        y={xAxisY}
+        height={cHeight}
+        width={cWidth}
+        margin={margin}
+        scale={xScale}
+        xGrid={xGrid}
+        type={xAxis}
+        label={xAxisLabel}
+        xTicksValue={xTicksValue}
+        />
+      )}
+        {layers.map((layer, i) => (
+          <path
+            key={i}
+            d={areaGenerator(layer)}
+            fill= {colorScale(layer.key)}
+          />
+        ))}
+    </g>
       </svg>
   );
 }
