+++ src/charts/LineChart/LineChart.tsx
@@ -4,7 +4,7 @@ import { useResponsive } from '../../hooks/useResponsive';
 import * as d3 from "d3";
 import Axis from "../../components/ContinuousAxis";
 import Line from '../../components/Line';
-import { LineChartProps, ColorScale, AccessorFunc, Data, GroupAccessorFunc } from "../../../types";
+import { LineChartProps, ColorScale, xAccessorFunc, yAccessorFunc, Data, GroupAccessorFunc } from "../../../types";
 import {
   getXAxisCoordinates,
   getYAxisCoordinates,
@@ -60,8 +60,8 @@ export default function LineChart({
     xDataType = inferXDataType(data[0], xKey);
   }
 
-  const xAccessor: AccessorFunc = xDataType === 'number' ? (d) => d[xKey] : (d) => new Date(d[xKey]);
-  const yAccessor: AccessorFunc = (d) => d[yKey];
+  const xAccessor: xAccessorFunc = xDataType === 'number' ? (d) => d[xKey] : (d) => new Date(d[xKey]);
+  const yAccessor: yAccessorFunc = (d) => d[yKey];
 
   const yScale = yScaleDef(data, yAccessor, margin, cHeight);
   const {xScale, xMin, xMax} = xScaleDef(data, xDataType, xAccessor, margin, cWidth, chart);
